Add merge native and soruce maps images test
